Rename REST module `routes` to `controller`
