+++ minerva-rest/src/controller/user.rs
